Moves functions from router to separate file
